Added functionality to isosurface permitting the optional output of fingerprints plots in matrix form. an example can be found in runfiles/fun_fingerprint.foo
Added method to vec{real} for getting the indices of all the values in a given range
Added method to vec{int} for getting the intersection of the vector with another vector

Uncommented appropriate methods for isosurface and molecule.base suboptions.

Added a filler method for vec{int}, allowing the populating of a newly created vector with
a single value.